+++ src/actions/record/RecordUpdate.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * @copyright  Copyright (c) Flipbox Digital Limited
+ * @license    https://github.com/flipboxfactory/craft-ember/blob/master/LICENSE
+ * @link       https://github.com/flipboxfactory/craft-ember
+ */
+
+namespace flipbox\ember\actions\record;
+
+use yii\base\Action;
+
+/**
+ * @author Flipbox Factory <hello@flipboxfactory.com>
+ * @since 1.0.0
+ */
+abstract class RecordUpdate extends Action
+{
+    use traits\Save, traits\Lookup;
+}